control-service: amazon ses uses real values.
