+++ projects/control-service/projects/pipelines_control_service/src/main/java/com/vmware/taurus/service/notification/DefaultEmailNotification.java
@@ -1,0 +1,117 @@
+/*
+ * Copyright 2023-2024 Broadcom
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package com.vmware.taurus.service.notification;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Component;
+
+import javax.mail.*;
+import javax.mail.internet.MimeMessage;
+import java.util.Arrays;
+import java.util.Properties;
+import java.util.stream.Collectors;
+
+/**
+ * The EmailNotification class is responsible for implementing the email notifications
+ * functionality. The single exposed send method accepts pre-build notification object.
+ *
+ * @see DataJobNotification
+ * @see NotificationContent
+ */
+@Component
+public class DefaultEmailNotification implements EmailNotification {
+
+  static Logger log = LoggerFactory.getLogger(DefaultEmailNotification.class);
+
+  private static final String CONTENT_TYPE = "text/html; charset=utf-8";
+
+  private final Session session;
+  private final EmailConfiguration emailConfiguration;
+
+  public DefaultEmailNotification(EmailConfiguration emailConfiguration) {
+    log.info("Constructing EmailNotification class");
+    this.emailConfiguration = emailConfiguration;
+    Properties properties = System.getProperties();
+    properties.putAll(emailConfiguration.smtpWithPrefix());
+    properties.setProperty("mail.transport.protocol", emailConfiguration.getTransportProtocol());
+    session = Session.getInstance(properties);
+  }
+
+  public void send(NotificationContent notificationContent) throws MessagingException {
+    log.info("Send EmailNotification");
+    if (recipientsExist(notificationContent.getRecipients())) {
+      try {
+        sendEmail(notificationContent, notificationContent.getRecipients());
+      } catch (SendFailedException firstException) {
+        log.info(
+            "Following addresses are invalid: {}",
+            concatAddresses(firstException.getInvalidAddresses()));
+        var validUnsentAddresses = firstException.getValidUnsentAddresses();
+        if (recipientsExist(validUnsentAddresses)) {
+          log.info("Retrying unsent valid addresses: {}", concatAddresses(validUnsentAddresses));
+          try {
+            sendEmail(notificationContent, validUnsentAddresses);
+          } catch (SendFailedException retriedException) {
+            log.warn(
+                "\nwhat: {}\nwhy: {}\nconsequence: {}\ncountermeasure: {}",
+                "Unable to send notification to the recipients",
+                retriedException.getMessage(),
+                "Listed recipients won't receive notifications for the job's status",
+                "There might be misconfiguration or outage. Check error message for more details");
+          }
+        }
+      }
+    }
+  }
+
+  private void sendEmail(NotificationContent notificationContent, Address[] recipients)
+      throws MessagingException {
+    if (emailConfiguration.isAuthEnabled()) {
+      sendAuthenticatedEmail(notificationContent, recipients);
+    } else {
+      sendUnauthenticatedEmail(notificationContent, recipients);
+    }
+  }
+
+  private void sendUnauthenticatedEmail(
+      NotificationContent notificationContent, Address[] recipients) throws MessagingException {
+    var mimeMessage = prepareMessage(notificationContent);
+    Transport.send(mimeMessage, recipients);
+  }
+
+  private void sendAuthenticatedEmail(NotificationContent notificationContent, Address[] recipients)
+      throws MessagingException {
+    Transport transport = session.getTransport();
+    var mimeMessage = prepareMessage(notificationContent);
+    try {
+      transport.connect(emailConfiguration.getUsername(), emailConfiguration.getPassword());
+      transport.sendMessage(mimeMessage, recipients);
+    } finally {
+      transport.close();
+    }
+  }
+
+  private MimeMessage prepareMessage(NotificationContent notificationContent)
+      throws MessagingException {
+    MimeMessage mimeMessage = new MimeMessage(session);
+    mimeMessage.setFrom(notificationContent.getSender());
+    mimeMessage.setRecipients(MimeMessage.RecipientType.TO, notificationContent.getRecipients());
+    mimeMessage.setSubject(notificationContent.getSubject());
+    mimeMessage.setContent(notificationContent.getContent(), CONTENT_TYPE);
+    return mimeMessage;
+  }
+
+  private boolean recipientsExist(Address[] recipients) {
+    return recipients.length > 0;
+  }
+
+  String concatAddresses(Address[] addresses) {
+    return addresses == null
+        ? null
+        : Arrays.stream(addresses).map(Address::toString).collect(Collectors.joining(" "));
+  }
+}
